requirement.txt
faiss-cpu==1.8.0
FastNLP==1.0.1
jieba==0.42.1
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.37
langchain-text-splitters==0.0.1
langdetect==1.0.9
nltk==3.8.1
numpy==1.26.4
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
tokenizers==0.13.2
torch==2.2.2
tqdm==4.66.2
transformers==4.26.1
zhipuai==2.0.1